Added the ability to add a post, though there's still many bugs (see description)
The like bool is global, so when you like 1 post, you'll dislike the other at the attempt of liking it.
You still get like 5 instances of 1 comment button being pressed for some reason.
Also there's the occasional disappearance of pictures, and appearance of blank images even when none is provided